# Optimize the matchLenSSE4 asm implementation.