--- a/flate/crc32_amd64.go
+++ b/flate/crc32_amd64.go
@@ -23,7 +23,9 @@ func crc32sseAll(a []byte, dst []hash)
 // matchLenSSE4 returns the number of matching bytes in a and b
 // up to length 'max'. Both slices must be at least 'max'
 // bytes in size.
-// It uses the PCMPESTRI SSE 4.2 instruction.
+//
+// TODO: drop the "SSE4" name, since it doesn't use any SSE instructions.
+//
 //go:noescape
 func matchLenSSE4(a, b []byte, max int) int