@nigeltao 2016-04-23
The same patch would apply to the snappy package, except #54 is the bigger issue.