@klauspost 2016-10-16
Finally got around to testing it thoroughly.  No issues found - thanks for the contribution.